refine: move mock transmitter code under lambda example
