+++ examples/aws-lambda/src/mock-transmitter/auth/cognitoService.ts
