+++ examples/aws-lambda/src/mock-transmitter/auth/handler.ts
