+++ examples/aws-lambda/src/mock-transmitter/auth/policy.ts
