+++ examples/aws-lambda/src/mock-transmitter/jwks/createJwksFromRawPublicKey.ts
