+++ examples/aws-lambda/src/mock-transmitter/jwks/handler.ts
