+++ examples/aws-lambda/src/mock-transmitter/kmsService.ts
