+++ examples/aws-lambda/src/mock-transmitter/responses.ts
