+++ examples/aws-lambda/src/mock-transmitter/types.ts
