+++ examples/aws-lambda/src/mock-transmitter/utils.ts
