+++ examples/aws-lambda/src/mock-transmitter/verification/constructVerificationSecurityEvent.ts
@@ -1,4 +1,4 @@
-import { SET, SETVerificationRequest } from './types'
+import { SET, SETVerificationRequest } from '../types'
 
 export function constructVerificationFullSecurityEvent(
   requestId: string,
