+++ examples/aws-lambda/src/mock-transmitter/verification/validation.ts
